processing/requirements.txt
kafka-python==2.0.2
redis==5.0.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-kafka-python==0.41b0
opentelemetry-instrumentation-redis==0.41b0
opentelemetry-exporter-otlp==1.20.0
psutil==5.9.5